requirements.txt
annotated-types==0.6.0
anyio==4.3.0
apturl==0.5.2
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asttokens==2.4.1
async-lru==2.0.4
attrs==23.2.0
Babel==2.14.0
bcrypt==3.2.0
beautifulsoup4==4.12.3
bleach==6.1.0
blinker==1.4
Brlapi==0.8.3
certifi==2020.6.20
cffi==1.16.0
chardet==4.0.0
charset-normalizer==3.3.2
click==8.0.3
colorama==0.4.4
comm==0.2.2
command-not-found==0.3
cryptography==3.4.8
cupshelpers==1.0
dbus-python==1.2.18
debugpy==1.8.1
decorator==5.1.1
defer==1.0.6
defusedxml==0.7.1
distlib==0.3.8
distro==1.7.0
distro-info==1.1+ubuntu0.1
duplicity==0.8.21
exceptiongroup==1.2.0
executing==2.0.1
fasteners==0.14.1
fastjsonschema==2.19.1
filelock==3.15.4
fqdn==1.5.1
future==0.18.2
gpg==1.16.0
h11==0.14.0
httpcore==1.0.4
httplib2==0.20.2
httpx==0.27.0
idna==3.3
importlib-metadata==4.6.4
iniconfig==2.0.0
ipykernel==6.29.3
ipython==8.22.2
ipython_genutils==0.2.0
isoduration==20.11.0
jedi==0.19.1
jeepney==0.7.1
Jinja2==3.1.3
json5==0.9.24
jsonpointer==2.4
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
jupyter-events==0.10.0
jupyter-lsp==2.2.4
jupyter_client==8.6.1
jupyter_core==5.7.2
jupyter_server==2.13.0
jupyter_server_terminals==0.5.3
jupyterlab==4.1.5
jupyterlab_pygments==0.3.0
jupyterlab_server==2.25.4
keyring==23.5.0
language-selector==0.1
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
lockfile==0.12.2
louis==3.20.0
macaroonbakery==1.3.1
Mako==1.1.3
MarkupSafe==2.0.1
matplotlib-inline==0.1.6
mistune==3.0.2
monotonic==1.6
more-itertools==8.10.0
MouseInfo==0.1.3
nbclient==0.10.0
nbconvert==7.16.2
nbformat==5.10.3
nest-asyncio==1.6.0
netifaces==0.11.0
notebook==7.1.2
notebook_shim==0.2.4
numpy==2.0.0
oauthlib==3.2.0
olefile==0.46
openai==1.25.0
overrides==7.7.0
packaging==24.0
pandocfilters==1.5.1
paramiko==2.9.3
parso==0.8.3
pexpect==4.8.0
pillow==10.3.0
platformdirs==4.2.0
pluggy==1.5.0
prometheus_client==0.20.0
prompt-toolkit==3.0.43
protobuf==3.12.4
psutil==5.9.8
ptyprocess==0.7.0
pure-eval==0.2.2
PyAutoGUI==0.9.54
pycairo==1.20.1
pycparser==2.21
pycups==2.0.1
pydantic==2.7.1
pydantic_core==2.18.2
pygame==2.5.2
PyGetWindow==0.0.9
Pygments==2.17.2
PyGObject==3.42.1
PyJWT==2.3.0
pymacaroons==0.13.0
PyMsgBox==1.0.9
PyNaCl==1.5.0
pyparsing==2.4.7
pyperclip==1.8.2
PyRect==0.2.0
pyRFC3339==1.1
PyScreeze==0.1.30
pytest==8.2.2
python-apt==2.4.0+ubuntu2
python-dateutil==2.9.0.post0
python-debian==0.1.43+ubuntu1.1
python-dotenv==0.19.2
python-json-logger==2.0.7
python3-xlib==0.15
pytweening==1.2.0
pytz==2022.1
pyxdg==0.27
PyYAML==5.4.1
pyzmq==25.1.2
referencing==0.34.0
reportlab==3.6.8
requests==2.31.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rpds-py==0.18.0
screen-resolution-extra==0.0.0
SecretStorage==3.3.1
Send2Trash==1.8.2
shapely==2.0.4
six==1.16.0
sniffio==1.3.1
soupsieve==2.5
ssh-import-id==5.11
stack-data==0.6.3
systemd-python==234
terminado==0.18.1
tinycss2==1.2.1
tk==0.1.0
tomli==2.0.1
tornado==6.4
tqdm==4.66.2
traitlets==5.14.2
types-python-dateutil==2.9.0.20240316
typing_extensions==4.10.0
ubuntu-advantage-tools==8001
ubuntu-drivers-common==0.0.0
ufw==0.36.1
unattended-upgrades==0.1
uri-template==1.3.0
urllib3==1.26.5
usb-creator==0.3.7
virtualenv==20.26.3
wadllib==1.3.6
wcwidth==0.2.13
webcolors==1.13
webencodings==0.5.1
websocket-client==1.7.0
xdg==5
xkit==0.0.0
zipp==1.0.0
